Add new test, remove initial test created. More test for section 2 here: https://app.notion.com/p/wpmedia/RocketCDN-Free-Tier-Integration-tests-planning-388ed22a22f0809cbf0ccfed59ea9cb6?source=copy_link#388ed22a22f08097ad9ffa8235c76293
